+++ R/internals.R
@@ -1,3 +1,4 @@
+
 #' Take the fingerprint of a data.frame - find the class or return NA
 #'
 #' `fingerprint` is an internal function that takes the "fingerprint" of a
@@ -7,18 +8,42 @@
 #'   which the package, `visdat`, is based upon
 #'
 #' @param x a vector
+#' @keywords internal
+#' @noRd
 #'
 fingerprint <- function(x){
 
   # is the data missing?
-  ifelse(is.na(x),
-         # yes? Leave as is NA
-         yes = NA,
-         # no? make that value no equal to the class of this cell.
-         no = glue::glue_collapse(class(x),
-                                  sep = "\n")
-  )
+  if (!is.list(x)) {
+    ifelse(is.na(x),
+           # yes? Leave as is NA
+           yes = NA,
+           # no? make that value no equal to the class of this cell.
+           no = glue::glue_collapse(class(x),
+                                    sep = "\n")
+    )
+  } else {
+    ifelse(purrr::map_lgl(x,~length(.x)==0),
+           # yes? Leave as is NA
+           yes = NA,
+           # no? make that value no equal to the class of this cell.
+           no = glue::glue_collapse(class(x),
+                                    sep = "\n")
+    )
+
+  }
 } # end function
+
+#' Run fingerprint on a dataframe
+#'
+#' @param x data frame
+#'
+#' @return data frame with column types and NA values
+#' @noRd
+#' @note internal
+fingerprint_df <- function(x){
+  purrr::map_df(x, fingerprint)
+}
 
 
 #' (Internal) Gather rows into a format appropriate for grid visualisation
@@ -27,13 +52,19 @@ fingerprint <- function(x){
 #'
 #' @return data.frame gathered to have columns "variables", "valueType", and a
 #'   row id called "rows".
+#' @keywords internal
+#' @noRd
 #'
 vis_gather_ <- function(x){
   x %>%
-  dplyr::mutate(rows = dplyr::row_number()) %>%
-    tidyr::gather_(key_col = "variable",
-                   value_col = "valueType",
-                   gather_cols = names(.)[-length(.)])
+    dplyr::mutate(rows = dplyr::row_number()) %>%
+    tidyr::pivot_longer(
+      cols = -rows,
+      names_to = "variable",
+      values_to = "valueType",
+      values_transform = list(valueType = as.character)
+    ) %>%
+    dplyr::arrange(rows, variable, valueType)
 }
 
 #' (Internal) Add values of each row as a column
@@ -45,15 +76,20 @@ vis_gather_ <- function(x){
 #' @param x dataframe created from `vis_gather_`
 #'
 #' @return the x dataframe with the added column `value`.
+#' @noRd
+#' @keywords internal
 #'
 vis_extract_value_ <- function(x){
 
-  suppressWarnings(
-    tidyr::gather_(x,
-                   "variable",
-                   "value",
-                   names(x))$value
+  data_longer <- tidyr::pivot_longer(
+    data = x,
+    cols = dplyr::everything(),
+    names_to = "variable",
+    values_to = "value",
+    values_transform = list(value = as.character)
   )
+
+  data_longer$value
 
 }
 
@@ -63,15 +99,17 @@ vis_extract_value_ <- function(x){
 #'   `vis_extract_value`.
 #'
 #' @return a ggplot object
+#' @keywords internal
+#' @noRd
 #'
 vis_create_ <- function(x){
 
   ggplot2::ggplot(data = x,
-                  ggplot2::aes_string(x = "variable",
-                                      y = "rows",
+                  ggplot2::aes(x = variable,
+                               y = rows,
                                     # text assists with plotly mouseover
-                                    text = "value")) +
-  ggplot2::geom_raster(ggplot2::aes_string(fill = "valueType")) +
+                                    text = value)) +
+  ggplot2::geom_raster(ggplot2::aes(fill = valueType)) +
   ggplot2::theme_minimal() +
   ggplot2::theme(axis.text.x = ggplot2::element_text(angle = 45,
                                                      vjust = 1,
@@ -80,7 +118,8 @@ vis_create_ <- function(x){
                 y = "Observations") +
     # flip the axes
     ggplot2::scale_y_reverse() +
-    ggplot2::theme(axis.text.x = ggplot2::element_text(hjust = 0.5))
+    ggplot2::theme(axis.text.x = ggplot2::element_text(hjust = 0.5)) +
+    ggplot2::guides(colour = "none")
 
 }
 
@@ -96,9 +135,9 @@ vis_create_ <- function(x){
 #'   http://colorbrewer2.org/.
 #'
 #' @return a visdat plot with a particular palette
-#'
+#' @keywords internal
+#' @noRd
 #' @examples
-#'
 #' \dontrun{
 #' # see internal use inside vis_guess and vis_dat
 #' }
@@ -106,74 +145,81 @@ vis_create_ <- function(x){
 add_vis_dat_pal <- function(vis_plot, palette){
 
   # palette options: http://docs.ggplot2.org/current/discrete_scale.html
-# qualitative, 6 colours --------------------------------------------------
-vis_pal_qual <- c("#e41a1c", # red
-                      "#ffff33", # yellow
-                      "#ff7f00", # Orange
-                      "#377eb8", # blue
-                      "#4daf4a", # Green
-                      "#984ea3") # Purple
+  # qualitative, 6 colours --------------------------------------------------
+  vis_pal_qual <- c("#e41a1c", # red
+                             "#ffff33", # yellow
+                             "#ff7f00", # Orange
+                             "#377eb8", # blue
+                             "#4daf4a", # Green
+                             "#984ea3") # Purple
 
-# diverging, 6 colours, colour-blind safe -------------------------------
-vis_pal_cb_safe <- c('#d73027', # red
-                         '#fc8d59', # orange
-                         '#fee090', # yellow
-                         '#e0f3f8', # light blue
-                         '#91bfdb', # mid blue
-                         '#4575b4') # dark blue
+  # diverging, 6 colours, colour-blind safe -------------------------------
+  vis_pal_cb_safe <- c('#d73027', # red
+                                '#fc8d59', # orange
+                                '#fee090', # yellow
+                                '#e0f3f8', # light blue
+                                '#91bfdb', # mid blue
+                                '#4575b4') # dark blue
 
-if (palette == "default"){
+  if (palette == "default"){
 
-  vis_plot
+    vis_plot
 
-} else if (palette == "qual") {
+  } else if (palette == "qual") {
 
-  vis_plot +
-    ggplot2::scale_fill_manual(limits = c("character",
-                                          "date",
-                                          "factor",
-                                          "integer",
-                                          "logical",
-                                          "numeric"),
-                               breaks = c("character", # red
-                                          "date", # orange
-                                          "factor", # yellow
-                                          "integer", # light blue
-                                          "logical", # mid blue
-                                          "numeric"), # dark blue
-                               values = vis_pal_qual,
-                               na.value = "grey")
+    vis_plot +
+      ggplot2::scale_fill_manual(limits = c("character",
+                                            "date",
+                                            "factor",
+                                            "integer",
+                                            "logical",
+                                            "numeric"),
+                                 breaks = c("character", # red
+                                            "date", # orange
+                                            "factor", # yellow
+                                            "integer", # light blue
+                                            "logical", # mid blue
+                                            "numeric"), # dark blue
+                                 values = vis_pal_qual,
+                                 na.value = "grey")
 
 
-} else if (palette == "cb_safe") {
+  } else if (palette == "cb_safe") {
 
-  vis_plot +
-    ggplot2::scale_fill_manual(limits = c("character",
-                                          "date",
-                                          "factor",
-                                          "integer",
-                                          "logical",
-                                          "numeric"),
-                               breaks = c("character", # red
-                                          "date", # orange
-                                          "factor", # yellow
-                                          "integer", # light blue
-                                          "logical", # mid blue
-                                          "numeric"), # dark blue
-                               values = vis_pal_cb_safe,
-                               na.value = "grey")
+    vis_plot +
+      ggplot2::scale_fill_manual(limits = c("character",
+                                            "date",
+                                            "factor",
+                                            "integer",
+                                            "logical",
+                                            "numeric"),
+                                 breaks = c("character", # red
+                                            "date", # orange
+                                            "factor", # yellow
+                                            "integer", # light blue
+                                            "logical", # mid blue
+                                            "numeric"), # dark blue
+                                 values = vis_pal_cb_safe,
+                                 na.value = "grey")
 
-} else  {
-  stop("palette arguments need to be either 'qual' 'cb_safe' or 'default'")
-} # close else brace
+  } else  {
+    cli::cli_abort(
+      c(
+        "Palette arguments need to be one of: 'qual', 'cb_safe', or 'default'",
+        "You palette argument was: {.arg {palette}}"
+      )
+    )
+  } # close else brace
 
 } # close the function
 
-#' (Internal) Create labels for the columns containing the \% missing data
+#' Create labels for the columns containing the percent missing data
 #'
 #' @param x data.frame
 #' @param col_order_index the order of the columns
-#'
+#' @note internal
+#' @keywords internal
+#' @noRd
 #' @return data.frame containing the missingness percent down to 0.1 percent
 #'
 label_col_missing_pct <- function(x,
@@ -181,17 +227,16 @@ label_col_missing_pct <- function(x,
 
   # present everything in the right order
 
-  labelled_pcts <-
-  purrr::map_df(x, ~round(mean(is.na(.))*100,2))[col_order_index] %>%
+  labelled_pcts <- colMeans(is.na(x))[col_order_index] %>%
     purrr::map_chr(function(x){
       dplyr::case_when(
         x == 0 ~  "0%",
-        x >= 0.1 ~ as.character(glue::glue("{x}%")),
-        x < 0.1 ~ "<0.1%"
+        x >= 0.001 ~ scales::percent(x, accuracy = 1),
+        x < 0.001 ~ "<0.1%"
       )
     })
 
-    glue::glue("{col_order_index} ({labelled_pcts})")
+  glue::glue("{col_order_index} ({labelled_pcts})")
 
 }
 
@@ -205,7 +250,8 @@ label_col_missing_pct <- function(x,
 #' @return a tibble with two columns `p_miss_lab` and `p_pres_lab`,
 #'   containing the labels to use for present and missing. A dataframe is
 #'   returned because I think it is a good style habit compared to a list.
-#'
+#' @noRd
+#' @keywords internal
 #'
 miss_guide_label <- function(x) {
 
@@ -234,7 +280,7 @@ miss_guide_label <- function(x) {
 
     p_pres_lab <- glue::glue("Present \n({p_pres}%)")
 
-    }
+  }
 
   label_frame <- tibble::tibble(p_miss_lab,
                                 p_pres_lab)
@@ -249,7 +295,10 @@ miss_guide_label <- function(x) {
 #' @param x data.frame
 #' @param ... optional extra inputs
 #'
-#' @return logical - TRUE means that there is a column with numerics, FALSE means that there is a column that is not numeric
+#' @return logical - TRUE means that there is a column with numerics, FALSE
+#'   means there is a column that is not numeric
+#' @noRd
+#' @keywords internal
 #'
 #' @examples
 #'
@@ -263,6 +312,11 @@ all_numeric <- function(x, ...){
 }
 # Can I capture moving from a value to NA, or, from NA to another value?
 
+is_binary <- function(x) all(x %in% c(0L, 1L, NA))
+
+all_binary <- function(x, ...){
+  all(as.logical(lapply(x, is_binary)))
+}
 
 #' Test if input is a data.frame
 #'
@@ -277,12 +331,112 @@ all_numeric <- function(x, ...){
 #' #fail
 #' test_if_dataframe(AirPassengers)
 #' }
-#'
+#' @keywords internal
+#' @noRd
 test_if_dataframe <- function(x){
+
   if (!inherits(x, "data.frame")) {
-    stop("vis_dat requires a data.frame but the object I see has class/es: ",
-         glue::glue_collapse(class(x),
-                             sep = ", "),
-         call. = FALSE)
+    cli::cli_abort(
+      c(
+        "{.code vis_dat()} requires a {.cls data.frame}",
+        "the object I see has class(es): ",
+        "{.cls {glue::glue_collapse(class(x), sep = ', ', last = ', and ')}}"
+      )
+    )
   }
+}
+
+test_if_all_numeric <- function(data){
+  if (!all_numeric(data)) {
+    cli::cli_abort(
+      c(
+        "Data input can only contain numeric values",
+        "Please subset the data to the numeric values you would like.",
+        "{.code dplyr::select(<data>, where(is.numeric))}",
+        "Can be helpful here!"
+      )
+    )
+  }
+}
+
+test_if_all_binary <- function(data){
+
+  if (!all_binary(data)) {
+    cli::cli_abort(
+      c(
+        "data input can only contain binary values",
+        "This means values are either 0 or 1, or NA.",
+        "Please subset the data to be binary values, or see {.code ?vis_value.}"
+      )
+    )
+  }
+}
+
+monotonic <- function(x) all(diff(x) == 0)
+
+#' Scale a vector between 0 and one.
+#'
+#' @param x numeric vector
+#'
+#' @return numeric vector between 0 and 1
+#' @noRd
+#' @keywords internal
+scale_01 <- function(x) {
+  if (monotonic(x)){
+    return((x/x))
+  }
+  (x - min(x, na.rm = TRUE)) / diff(range(x, na.rm = TRUE))
+}
+
+group_by_fun <- function(data,.fun, ...){
+  tidyr::nest(data) %>%
+    dplyr::mutate(data = purrr::map(data, .fun, ...)) %>%
+    tidyr::unnest(cols = c(data))
+}
+
+data_vis_class_not_implemented <- function(fun){
+  cli::cli_abort(
+    c(
+      "We have not (yet) implemented the method for {.code fun} for \\
+      object with class(es): ",
+      "{.cls {glue::glue_collapse(class(x), sep = ', ', last = ', and ')}}"
+    )
+  )
+}
+
+update_col_order_index <- function(col_order_index, facet, env = environment()){
+  group_string <- deparse(substitute(facet, env = env))
+
+  facet_position <- which(col_order_index == group_string)
+  col_order_index <- col_order_index[-facet_position]
+}
+
+test_if_large_data <- function(x, large_data_size, warn_large_data){
+  if (ncol(x) * nrow(x) > large_data_size && warn_large_data){
+    cli::cli_abort(
+      c(
+        "Data exceeds recommended size for visualisation",
+        "Consider downsampling your data with {.fn dplyr::slice_sample}",
+        "Or set argument, {.arg warn_large_data} = {.arg FALSE}"
+      )
+    )
+  }
+}
+
+fast_n_miss_col <- function(x) colSums(is.na(x))
+
+n_miss_col <- function(data, sort = FALSE){
+  # if no list columns
+  any_list <- any(purrr::map_lgl(data, is.list))
+  if (!any_list){
+    n_missing_cols <- fast_n_miss_col(data)
+  } else if (any_list){
+    n_missing_cols <- fast_n_miss_col(fingerprint_df(data))
+  }
+
+ if (sort){
+   n_missing_cols <- sort(n_missing_cols, decreasing = TRUE)
+ }
+
+  n_missing_cols
 }